Given a 2d grid map of '1's (land) and '0's (water), count the number of islands. An island is surrounded by water and is formed by connecting adjacent lands horizontally or vertically. You may assume all four edges of the grid are all surrounded by water.
A graph where all vertices are connected with each other has exactly one connected component, consisting of the whole graph. Such a graph with only one connected component is called a Strongly Connected Graph.

The problem can be easily solved by applying DFS() on each component. In each DFS() call, a component or a sub-graph is visited. We will call DFS on the next un-visited component. The number of calls to DFS() gives the number of connected components. BFS can also be used.

What is an island?
A group of connected 1s forms an island. For example, the below matrix contains 4 islandsInput 

i am marking similar group as X ,Y,Z,A,B
                  {{1X, 1X, 0, 0, 0},
                   {0, 1X, 0, 0, 1Y},
                   {1X, 0, 0, 1Y, 1Y},
                   {0, 0, 0, 0, 0},
                   {1Z, 0, 1A, 0, 1B} }
                 Output : 5
here  if we talk about output that why output is 5 , then firstly male pair of 1's in matrix as you see in above example X is first group , Y pair is second group, hence Z has no pair so it is singly a group and similarly A and B
So total 5 output comes