LeetcodeProblemno-200Solution-https://leetcode.com/problems/number-of-islands/
A graph where all vertices are connected with each other has exactly one connected component, consisting of the whole graph. Such a graph with only one connected component is called a Strongly Connected Graph.

The problem can be easily solved by applying DFS() on each component. In each DFS() call, a component or a sub-graph is visited. We will call DFS on the next un-visited component. The number of calls to DFS() gives the number of connected components. BFS can also be used.
What is an island?
A group of connected 1s forms an island. For example, the below matrix contains 4 islands
                  {{1X, 1X, 0, 0, 0},
                   {0, 1X, 0, 0, 1Y},
                   {1X, 0, 0, 1Y, 1Y},
                   {0, 0, 0, 0, 0},
                   {1Z, 0, 1A, 0, 1B} 

                   Output : 5
here if we mark a group of pair as same alpahabet let us say X,Y,Z,A,B .
in first pair X 
in 2nd pair Y
now Z is alone so it is a group itself
and similary for A
similary for B
So output is 5.